+++ pyprom/lib/containers/inverse_edgepoint.py
@@ -7,13 +7,11 @@ the LICENSE file that accompanies it.
 This library contains a base container class for storing GridPoint
 type location objects.
 """
-
 from collections import defaultdict
-from .base import _Base
-from .shore import ShoreContainer
+from .gridpoint import GridPointContainer
 
 
-class InverseEdgePointContainer(_Base):
+class InverseEdgePointContainer(object):
     """
     Container for :class:`InverseEdgePoint` type lists.
     Allows for various list transformations.
@@ -33,9 +31,7 @@ class InverseEdgePointContainer(_Base):
         if inverseEdgePointList:
             self.points = inverseEdgePointList
         self.datamap = datamap
-        self.exemptPoints = defaultdict(list)
         self.mapEdge = mapEdge
-        self.branchMapEdge = False
 
     def iterNeighborDiagonal(self, inverseEdgePoint):
         """
@@ -73,155 +69,38 @@ class InverseEdgePointContainer(_Base):
             else:
                 continue
 
-    def findLinear(self):
+    def findHighEdges(self, elevation):
         """
-        :return: list of GridPoint containers representing individual edges
+        Hopefully a way more efficient way of finding high edges.
+        :return:
         """
-
-        shoreContainers = list()
-        self.exemptPoints = defaultdict(list)
-
-        rounds = 0
-        self.edge = list()
-        two = list()
-
-        # Find points with exactly one neighbor. These are edgepoints.
-        # Subdivide these into two groups, edges and stubs.
-
-        for point in (pt for pt in self.points):
-            neighbors = [x for x in self.iterNeighborOrthogonal(point)]
-            if len(neighbors) == 1:
-                if point.x in [self.datamap.max_x, 0] or\
-                                point.y in [self.datamap.max_y, 0]:
-                    self.edge.append(point)
-            if len(neighbors) == 2:
-                two.append(point)
-
-        # order the points
-        scanOrder = self.edge + two + self.points
-
-        while True:
-            rounds += 1
-            for point in (pt for pt in scanOrder
-                          if pt.y not in self.exemptPoints[pt.x]):
-                neighbors = [x for x in self.iterNeighborOrthogonal(point)]
-                masterPoint = point
-
-                if not len(neighbors):
-                    self.exemptPoints[point.x].append(point.y)
-                    if point.x in (0, self.datamap.max_x) or \
-                       point.y in (0, self.datamap.max_y):
-                        shoreContainers.append(ShoreContainer([point], True))
-                    else:
-                        shoreContainers.append(ShoreContainer([point]))
-                    break
-
-                firstPoint = neighbors[0]
-                shoreContainers.append(ShoreContainer(
-                    self.branchChaser(masterPoint,
-                                      masterPoint,
-                                      firstPoint), self.branchMapEdge))
-                self.branchMapEdge = False
-                break
-
-            if not len([pt for pt in self.points if pt.y not in
-                        self.exemptPoints[pt.x]]):
-                return shoreContainers
-            if rounds > 100:
-                self.logger.info('Something broke in {}'.format(self))
-                return shoreContainers
-
-    def branchChaser(self, masterPoint, originalPoint, firstPoint):
-        """
-        Recursive function for chasing down inverse edge Branches
-        :param masterPoint: Master point for this segment. This is
-         passed all the way through.
-        :param originalPoint: Original Point Branch under analysis.
-        :param firstPoint: First point, this is a neighbor of the
-         OriginalPoint and determines the direction of analysis travel.
-        :return: ordered List of points.
-        """
-
-        lookbackPoint = originalPoint
-        currentPoint = firstPoint
-
-        orderedList = [originalPoint]
-
-        while True:
-            # First, we find all neighbors who are not the original point,
-            # a lookback point, the master point, or an already analyzed point.
-
-            neighbors = [pt for pt in self.iterNeighborOrthogonal(currentPoint)
-                         if pt not in [lookbackPoint,
-                                       originalPoint,
-                                       masterPoint]]
-            neighbors = [pt for pt in neighbors
-                         if pt.y not in self.exemptPoints[pt.x]]
-
-            # More than one neighbor? We'll need to look back at the last
-            # point and find how common neighbors we have.
-            commonEdgeHash = defaultdict(list)
-            if currentPoint.x in (0, self.datamap.max_x) or\
-               currentPoint.y in (0, self.datamap.max_y):
-                self.branchMapEdge = True
-            if len(neighbors) > 1:
-                for neighbor in neighbors:
-                    commonEdgePoints =\
-                        len(set(neighbor.edgePoints).
-                            intersection(lookbackPoint.edgePoints))
-                    commonEdgeHash[commonEdgePoints].append(neighbor)
-
-                # Look for the neighbors with the most EdgePoints in common
-                if len(commonEdgeHash[max(commonEdgeHash.keys())]) != 1:
-                    level1CommonEdgeHash = defaultdict(list)
-                    # a bunch of neighbors with no common EdgePoints? Lets
-                    # look for common edges with the neighbors
-                    if max(commonEdgeHash.keys()) == 0:
-                        for nei in commonEdgeHash[0]:
-                            level1CommonEdgePoints = len(nei.edgePoints)
-                            level1CommonEdgeHash[level1CommonEdgePoints].\
-                                append(nei)
-
-                        if len(level1CommonEdgeHash[max(
-                                level1CommonEdgeHash.keys())]) != 1:
-                            # Fucking hell, they're equal. Guess we'll just
-                            # choose the first one anyways.
-                            self.logger.debug("Highly unusual Branch "
-                                              "Exception! on {}".format(self))
-                            self.logger.debug(
-                                "TroubleMakers {}".format(currentPoint))
-
-                        orderedList +=\
-                            self.branchChaser(
-                                masterPoint,
-                                currentPoint,
-                                level1CommonEdgeHash[max(
-                                    level1CommonEdgeHash.keys())][0])
-                        continue
-                else:
-                    # Follow the branch with the most common neighbors.
-                    orderedList +=\
-                        self.branchChaser(
-                            masterPoint,
-                            currentPoint,
-                            commonEdgeHash[max(commonEdgeHash.keys())][0])
-                    continue
-
-            # Not exempt? Add to the ordered list.
-            if currentPoint.y not in self.exemptPoints[currentPoint.x]:
-                # Last modification in PR, delete this comment someday.
-                self.exemptPoints[currentPoint.x].append(currentPoint.y)
-                orderedList.append(currentPoint)
-
-            if len(neighbors) == 0:
-                # End of the line? return the ordered list.
-                for point in orderedList:
-                    self.exemptPoints[point.x].append(point.y)
-                return orderedList
-
-            # Just one neighbor? Okay, do this...
-            lookbackPoint = currentPoint
-            currentPoint = neighbors[0]
+        purgedIndex = defaultdict(list)
+        highLists = list()
+        for point in self.points:
+            if point.y in purgedIndex[point.x]:
+                continue
+            if point.elevation > elevation:
+                toBeAnalyzed = [point]
+                highList = list()
+                while True:
+                    if not toBeAnalyzed:
+                        highLists.append(highList)
+                        break
+                    try:
+                        gridPoint = toBeAnalyzed.pop()
+                    except:
+                        highLists.append(highList)
+                        break
+                    purgedIndex[gridPoint.x].append(gridPoint.y)
+                    highList.append(gridPoint)
+                    neighbors = [x for x in
+                                 self.iterNeighborOrthogonal(gridPoint)
+                                 if x.elevation > elevation and
+                                 x.y not in purgedIndex[x.x]]
+                    toBeAnalyzed += neighbors
+            else:
+                purgedIndex[point.x].append(point.y)
+        return [GridPointContainer(x) for x in highLists]
 
     def __repr__(self):
         return "<InverseEdgePointContainer>" \