+++ pyprom/lib/locations/spot_elevation.py
@@ -8,7 +8,6 @@ This library contains a base class for Coordinate oriented objects with
 Elevation data.
 """
 import json
-import gridpoint as g
 from .base_coordinate import BaseCoordinate
 
 
@@ -50,10 +49,11 @@ class SpotElevation(BaseCoordinate):
         :param datamap: :class:`Datamap` object
         :return: :class:`GridPoint object`
         """
-        return g.GridPoint(datamap.relative_position_latitude(self.latitude),
-                           datamap.relative_position_longitude(
+        from .gridpoint import GridPoint
+        return GridPoint(datamap.relative_position_latitude(self.latitude),
+                         datamap.relative_position_longitude(
                             self.longitude),
-                           self.elevation)
+                         self.elevation)
 
     @property
     def feet(self):