+++ pyprom/logic.py
@@ -11,14 +11,14 @@ import numpy
 import logging
 
 from collections import defaultdict
+from timeit import default_timer
+from datetime import timedelta
 from lib.locations.gridpoint import GridPoint
 from lib.locations.saddle import Saddle
 from lib.locations.summit import Summit
 from lib.locations.inverse_edgepoint import InverseEdgePoint
-from lib.locations.edgepoint import EdgePoint
 from lib.containers.spot_elevation import SpotElevationContainer
 from lib.containers.multipoint import MultiPoint
-from lib.containers.edgepoint import EdgePointContainer
 from lib.containers.inverse_edgepoint import InverseEdgePointContainer
 from lib.containers.high_edge import HighEdgeContainer
 from lib.containers.gridpoint import GridPointContainer
@@ -40,9 +40,7 @@ class AnalyzeData(object):
         self.max_x = self.datamap.max_x
         self.span_latitude = self.datamap.span_latitude
         self.cardinalGrid = dict()
-        self.skipSummitAnalysis = defaultdict(list)
-        self.summitObjects = SpotElevationContainer([])
-        self.saddleObjects = SpotElevationContainer([])
+        self.skipAnalysis = defaultdict(list)
 
     def analyze(self):
         """
@@ -50,7 +48,11 @@ class AnalyzeData(object):
         Looks for :class:`Summit`s, and :class:`Saddle`s
         return: (:class:`SpotElevationContainer`,SpotElevationContainer)
         """
+        self.start = default_timer()
+        self.lasttime = self.start
         self.logger.info("Initiating Analysis")
+        self.summitObjects = SpotElevationContainer([])
+        self.saddleObjects = SpotElevationContainer([])
         iterator = numpy.nditer(self.data, flags=['multi_index'])
         index = 0
         # Iterate through numpy grid, and keep track of gridpoint coordinates.
@@ -61,18 +63,70 @@ class AnalyzeData(object):
             # Quick Progress Meter. Needs refinement,
             index += 1
             if not index % 100000:
-                self.logger.info("{}/{} - {}%".format(index, self.data.size,
-                                 (index/self.data.size)*100))
 
-            # Check for summit
-            self._summit_and_saddle(x, y)
+                thisTime = default_timer()
+                split = round(thisTime - self.lasttime, 2)
+                self.lasttime = default_timer()
+                rt = self.lasttime - self.start
+                pointsPerSec = round(index/rt, 2)
+                self.logger.info(
+                    "Points per second: {} - {}%"
+                    " runtime: {}, split: {}".format(
+                        pointsPerSec,
+                        round(index/self.data.size * 100, 2),
+                        (str(timedelta(seconds=round(rt, 2)))),
+                        split
+                    ))
+
+            # Check for summit or saddle
+            result = self.summit_and_saddle(x, y)
+            if result:
+                if isinstance(result, Saddle):
+                    self.saddleObjects.points.append(result)
+                if isinstance(result, Summit):
+                    self.summitObjects.points.append(result)
             # Reset variables, and go to next gridpoint.
             self.edge = False
             self.blob = None
             iterator.iternext()
+        # Free some memory.
+        del(self.skipAnalysis)
         return self.summitObjects, self.saddleObjects
 
-    def _summit_and_saddle(self, x, y):
+    def analyze_multipoint(self, x, y, ptElevation):
+        """
+        :param x:
+        :param y:
+        :param ptElevation: Elevation of Multipoint Blob
+        :return: Summit, Saddle, or None
+        """
+        self.blob = self.equalHeightBlob(x, y, ptElevation)
+        highInverseEdge = self.blob.inverseEdgePoints.findHighEdges(
+            self.elevation)
+
+        for exemptPoint in self.blob.points:
+            self.skipAnalysis[exemptPoint.x] \
+                .append(exemptPoint.y)
+        if not len(highInverseEdge):
+            summit = Summit(self.datamap.x_position_latitude(x),
+                            self.datamap.y_position_longitude(y),
+                            self.elevation,
+                            edge=self.edge,
+                            multiPoint=self.blob
+                            )
+            return summit
+        if (len(highInverseEdge) > 1) or\
+                (len(highInverseEdge) == 1 and self.edge):
+            saddle = Saddle(self.datamap.x_position_latitude(x),
+                            self.datamap.y_position_longitude(y),
+                            self.elevation,
+                            edge=self.edge,
+                            multiPoint=self.blob,
+                            highShores=highInverseEdge)
+            return saddle
+        return None
+
+    def summit_and_saddle(self, x, y):
         """
         :param x:
         :param y:
@@ -80,68 +134,11 @@ class AnalyzeData(object):
         """
 
         # Exempt! bail out!
-        if y in self.skipSummitAnalysis[x]:
-            return
+        if y in self.skipAnalysis[x]:
+            return None
 
         saddleProfile = ["HLHL", "LHLH"]
         summitProfile = "L"
-
-        def _analyze_multipoint(x, y, ptElevation):
-            self.blob = self.equalHeightBlob(x, y, ptElevation)
-            pseudoShores = self.blob.inverseEdgePoints.findLinear()
-            summitLike = False
-
-            # Go find the shore of each blob, and assign a "H"
-            # for points higher than the equalHeightBlob, and "L"
-            # for points lower.
-            for shoreSet in pseudoShores:
-                shoreProfile = ""
-                for shorePoint in shoreSet.points:
-                    if not shorePoint.elevation:
-                        continue
-                    if shorePoint.elevation > ptElevation:
-                        shoreProfile += "H"
-                    if shorePoint.elevation < ptElevation:
-                        shoreProfile += "L"
-                reducedNeighborProfile = compressRepetetiveChars(shoreProfile)
-
-                if any(x in reducedNeighborProfile for x in saddleProfile)\
-                        and not summitLike:
-                    for exemptPoint in self.blob.points:
-                        self.skipSummitAnalysis[exemptPoint.x] \
-                            .append(exemptPoint.y)
-                    shores = HighEdgeContainer(shoreSet, ptElevation)
-                    saddle = Saddle(self.datamap.x_position_latitude(x),
-                                    self.datamap.y_position_longitude(y),
-                                    self.elevation,
-                                    edge=self.edge,
-                                    multiPoint=self.blob,
-                                    highShores=shores)
-                    self.saddleObjects.points.append(saddle)
-                    return
-                elif reducedNeighborProfile == summitProfile:
-                    summitLike = True
-                else:
-                    summitLike = False
-                    break
-
-            if summitLike:
-                for exemptPoint in self.blob.points:
-                    self.skipSummitAnalysis[exemptPoint.x] \
-                        .append(exemptPoint.y)
-                summit = Summit(self.datamap.x_position_latitude(x),
-                                self.datamap.y_position_longitude(y),
-                                self.elevation,
-                                edge=self.edge,
-                                multiPoint=self.blob
-                                )
-                self.summitObjects.points.append(summit)
-                return
-            else:
-                for exemptPoint in self.blob.points:
-                    self.skipSummitAnalysis[exemptPoint.x] \
-                        .append(exemptPoint.y)
-                return
 
         # Label this as an mapEdge under the following condition
         if x in (self.max_x, 0) or y in (self.max_y, 0):
@@ -157,10 +154,9 @@ class AnalyzeData(object):
             # a special MultiPoint analysis function.
             if not elevation:
                 continue
-            if elevation == self.elevation and _y not in\
-                            self.skipSummitAnalysis[_x]:
-                _analyze_multipoint(_x, _y, elevation)
-                return
+            if elevation == self.elevation and _y not in \
+                    self.skipAnalysis[_x]:
+                return self.analyze_multipoint(_x, _y, elevation)
             if elevation > self.elevation:
                 neighborProfile += "H"
             if elevation < self.elevation:
@@ -173,7 +169,7 @@ class AnalyzeData(object):
                             self.datamap.y_position_longitude(y),
                             self.elevation,
                             edge=self.edge)
-            self.summitObjects.points.append(summit)
+            return summit
 
         elif any(x in reducedNeighborProfile for x in saddleProfile):
             shores = HighEdgeContainer(shoreSet, self.elevation)
@@ -181,9 +177,9 @@ class AnalyzeData(object):
                             self.datamap.y_position_longitude(y),
                             self.elevation,
                             edge=self.edge,
-                            highShores=shores)
-            self.saddleObjects.points.append(saddle)
-        return
+                            highShores=shores.highPoints)
+            return saddle
+        return None
 
     def equalHeightBlob(self, x, y, elevation):
         """
@@ -198,27 +194,15 @@ class AnalyzeData(object):
         equalHeightHash = defaultdict(list)
         equalHeightHash[x].append(y)
         nesteddict = lambda: defaultdict(nesteddict)
-        edgeHash = nesteddict()  # {X : { Y : EdgePoint}}
         inverseEdgeHash = nesteddict()  # InverseEdgepoint (shore).
         toBeAnalyzed = [masterGridPoint]
-
-        # Helper function for equal neighbors.
-        def addEqual():
-            if edgeHash[gridPoint.x][gridPoint.y]:
-                    edgeHash[gridPoint.x][gridPoint.y]. \
-                         equalNeighbors.append(branch)
-            # Does not exist? Create.
-            else:
-                edgeHash[gridPoint.x][gridPoint.y] = \
-                    EdgePoint(gridPoint.x, gridPoint.y,
-                              gridPoint.elevation,
-                              [], [branch])
 
         # Loop until pool of equalHeight neighbors has been exhausted.
         while toBeAnalyzed:
             gridPoint = toBeAnalyzed.pop()
             neighbors = self.datamap.iterateDiagonal(gridPoint.x, gridPoint.y)
-            if gridPoint.x in (self.max_x, 0) or gridPoint.y in\
+            # Determine if edge or not.
+            if gridPoint.x in (self.max_x, 0) or gridPoint.y in \
                     (self.max_y, 0):
                 self.edge = True
             for _x, _y, elevation in neighbors:
@@ -227,120 +211,14 @@ class AnalyzeData(object):
                     branch = GridPoint(_x, _y, elevation)
                     equalHeightHash[_x].append(_y)
                     toBeAnalyzed.append(branch)
-                    addEqual()
-                # Equal and exempt? add to equal neighbor list.
-                elif elevation == gridPoint.elevation:
-                    addEqual()
-                elif elevation != gridPoint.elevation:
-                    # EdgePoint Object Exists? append nonEqual
-                    if edgeHash[gridPoint.x][gridPoint.y]:
-                        edgeHash[gridPoint.x][gridPoint.y]. \
-                            nonEqualNeighbors.append(GridPoint(
-                                 _x, _y, elevation))
-                    # Does not exist? Create.
-                    else:
-                        edgeHash[gridPoint.x][gridPoint.y] = \
-                            EdgePoint(gridPoint.x, gridPoint.y,
-                                      gridPoint.elevation,
-                                      [GridPoint(_x, _y, elevation)], [])
-
-                    # Add inverse EdgePoints (aka shores).
-                    if inverseEdgeHash[_x][_y]:
-                        inverseEdgeHash[_x][_y].addEdge(
-                            edgeHash[gridPoint.x][gridPoint.y])
-                    else:
+                elif elevation != masterGridPoint.elevation:
+                    if not inverseEdgeHash[_x][_y]:
                         inverseEdgeHash[_x][_y] = \
-                            InverseEdgePoint(_x, _y, elevation,
-                                             [edgeHash[gridPoint.x]
-                                              [gridPoint.y]])
+                            InverseEdgePoint(_x, _y, elevation)
 
         return MultiPoint(coordinateHashToGridPointList(equalHeightHash),
                           masterGridPoint.elevation, self.datamap,
-                          edgePoints=EdgePointContainer(
-                              edgePointIndex=edgeHash),
                           inverseEdgePoints=InverseEdgePointContainer(
                               inverseEdgePointIndex=inverseEdgeHash,
                               datamap=self.datamap, mapEdge=self.edge)
                           )
-
-
-class EqualHeightBlob(object):
-    """
-    I'm really just keeping this around for testing.
-    """
-    def __init__(self, x, y, elevation, datamap):
-        self.datamap = datamap
-        self.gridPoint = GridPoint(x, y, elevation)
-        self.equalHeightBlob = list()  # [[x,y]]
-        self.equalHeightHash = defaultdict(list)
-        self.equalHeightHash[x].append(y)
-
-        nesteddict = lambda: defaultdict(nesteddict)
-        self.edgeHash = nesteddict()  # {X : { Y : EdgePoint}}
-        self.inverseEdgeHash = nesteddict()  # Inverse Edgepoint (shore).
-        self.edge = False
-        self.buildBlob([self.gridPoint])
-
-    def buildBlob(self, toBeAnalyzed):
-
-        def addEqual(branch):
-            if self.edgeHash[gridPoint.x][gridPoint.y]:
-                self.edgeHash[gridPoint.x][gridPoint.y]. \
-                    equalNeighbors.append(branch)
-            # Does not exist? Create.
-            else:
-                self.edgeHash[gridPoint.x][gridPoint.y] = \
-                    EdgePoint(gridPoint.x, gridPoint.y,
-                              gridPoint.elevation,
-                              [], [branch])
-
-        while toBeAnalyzed:
-            gridPoint = toBeAnalyzed.pop()
-            neighbors = self.datamap.iterateDiagonal(gridPoint.x, gridPoint.y)
-            for _x, _y, elevation in neighbors:
-                branch = GridPoint(_x, _y, elevation)
-                if elevation == self.gridPoint.elevation and _y not in\
-                                    self.equalHeightHash[_x]:
-                    self.equalHeightHash[_x].append(_y)
-                    toBeAnalyzed.append(branch)
-                    if _x in (self.datamap.max_x, 0) or _y in\
-                            (self.datamap.max_y, 0):
-                        self.edge = True
-                    addEqual(branch)
-                elif elevation == self.gridPoint.elevation:
-                    addEqual(branch)
-                # Non Equal?
-                elif elevation != self.gridPoint.elevation:
-                    # EdgePoint Object Exists? append nonEqual
-                    if self.edgeHash[gridPoint.x][gridPoint.y]:
-                        self.edgeHash[gridPoint.x][gridPoint.y].\
-                            nonEqualNeighbors.append(branch)
-                    # Does not exist? Create.
-                    else:
-                        self.edgeHash[gridPoint.x][gridPoint.y] = \
-                            EdgePoint(gridPoint.x, gridPoint.y,
-                                      gridPoint.elevation,
-                                      [branch], [])
-
-                    # Add inverse EdgePoints (aka shores).
-                    if self.inverseEdgeHash[_x][_y]:
-                        self.inverseEdgeHash[_x][_y].addEdge(
-                            self.edgeHash[gridPoint.x][gridPoint.y])
-                    else:
-                        self.inverseEdgeHash[_x][_y] = \
-                            InverseEdgePoint(_x, _y, elevation,
-                                [self.edgeHash[gridPoint.x][gridPoint.y]])
-
-        self.equalHeightBlob =\
-            MultiPoint(coordinateHashToGridPointList(
-                       self.equalHeightHash),
-                       self.gridPoint.elevation,
-                       self.datamap,
-                       edgePoints=
-                       EdgePointContainer(edgePointIndex=self.edgeHash),
-                       inverseEdgePoints=
-                       InverseEdgePointContainer(inverseEdgePointIndex=
-                                                 self.inverseEdgeHash,
-                                                 datamap=self.datamap,
-                                                 mapEdge=self.edge)
-                       )